Implemented isAchievable method for workflows.

+++ src/main/java/metrics/EnvironmentNoise.java
@@ -8,12 +8,12 @@ public class EnvironmentNoise extends Metric {
 	}
 
 	@Override
-	public double serialComposition(double value1, double value2) {
+	public float serialComposition(float value1, float value2) {
 		return Math.max(value1, value2);
 	}
 
 	@Override
-	public double parallelComposition(double value1, double value2) {
+	public float parallelComposition(float value1, float value2) {
 		return Math.max(value1, value2);
 	}
 }
