+++ src/main/java/metrics/FalseNegativePercentage.java
@@ -7,12 +7,12 @@ public class FalseNegativePercentage extends metrics.Metric{
 		name="FalseNegative";
 	}
 	@Override
-	public double serialComposition(double value1, double value2) {
+	public float serialComposition(float value1, float value2) {
 		return Math.max(value1, value2);
 	}
 
 	@Override
-	public double parallelComposition(double value1, double value2) {
+	public float parallelComposition(float value1, float value2) {
 		return Math.max(value1, value2);
 	}
 }
